requirements.txt
###### Requirements without Version Specifiers ######
selenium
pylint
load_dotenv
python-dotenv